+++ dst/jobs/triggered/importEvents/run.js
@@ -123,9 +123,9 @@ function main(connection) {
         // 特定のコードの券種しか取り込まない
         const searchTicketTypesResult = yield offerService.searchTicketTypes({
             limit: 100,
-            project: { ids: [project.id] },
-            ids: offerCatalog.itemListElement.map((element) => element.id),
-            identifiers: offerCodes
+            project: { id: { $eq: project.id } },
+            id: { $in: offerCatalog.itemListElement.map((element) => element.id) },
+            identifier: { $in: offerCodes }
         });
         const ticketTypes = searchTicketTypesResult.data;
         debug(ticketTypes.length, 'ticketTypes found');