+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - イベント情報に取り込むオファーコードを限定するように設定追加
+- update @chevre/api-nodejs-client
 
 ### Deprecated