+++ CHANGELOG.md
@@ -10,9 +10,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
-- イベント情報に取り込むオファーコードを限定するように設定追加
-- update @chevre/api-nodejs-client
-
 ### Deprecated
 
 ### Removed
@@ -20,6 +17,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Fixed
 
 ### Security
+
+## v33.2.0 - 2020-03-10
+
+### Changed
+
+- イベント情報に取り込むオファーコードを限定するように設定追加
+- update @chevre/api-nodejs-client
 
 ## v33.1.0 - 2020-03-07